[OM-64079] Replaced collector account key for telemetry in t8c-operator.
This was done because the old one was accidentally published to GitHub.

Collectors are write-only accounts so there is no risk of someone with
this key reading customer telemetry from our customers, and it is
possible for a determined attacker to get this key in other ways but we
still shouldn't publish it.

Currently we haven't launched telemetry yet so invalidating the old key
won't cause any disruption.

The process to invalidate and replace a key is as follows:

1. Log into DataCloud and go to [Administration/Collector Accounts tab of the telemetry production project](https://datacloud.turbonomic.io/administration/collectoraccounts?projectId=turbo-telemetry-prod).
2. Press "Delete", then "New Account".
3. Save the `collector.json` file when prompted.
4. Run the command `base64 collector.json` from within the directory
   where you saved the file in step #3.
5. Copy the output to `xl/values.yaml` as seen in this commit.

Reviewed at https://rbcommons.com/s/VMTurbo/r/42413/