[OM-64251]: RB:42516 Fix translation for plan container VCPU actions
** Problem **

After the implementation of VCPU action translation for containers to show capacity values in millicores instead of MHz (OM_63851), VCPU resize actions for containers are no longer showing in plans.

This is because translation for these actions requires the presence of a TopologyGraph in the ActionOrchestrator, but the ActionOrchestrator does not keep a TopologyGraph for plans. As a result, all container VCPU actions for plans wound up with a failed translation and were not being shown.

** Solution **

Continue to translate container VCPU actions using the topology graph when it is available, but when it is not available, perform the translation by fetching the required data from the repository. This is similar to the strategy used for translating VM CPU actions (to go from MHz to core).

** Testing Done **
* Ensure that realtime container VCPU actions are displayed and translated correctly.
* Ensure that plan container VCPU actions now show up and are translated correctly.
* Add new unit test for the case when we have to fetch translation information from the repository for container VCPU action translation.