[OM-64327]: RB-42532 Fix for target update action.
During the update check if the target with the same identifier have ever existed. If we find that this target was created and exists right now, then throw an exception. If we find the target was created but now doesn't exist then remove identity information because it should be updated.