[OM-63845]: RB-42565: Add FAILING state when POST runs after a failed IN_PROGRESS
Intent
======
Add FAILING state when POST runs after a failed IN_PROGRESS so that POST workflows know if the IN_PROGRESS step failed.

Problem
=======
Services is unable to quarentine VMs using POST action scripts at TD Bank.

Diagnosis
=========
In classic, we update the action as failed before it reach the POST workflow. That's not the case in XL. In XL, when the action reaches the POST script, it's still IN_PROGRESS. This makes it impossible for Services to quarantine failing VMs in the POST action script because they have no idea if the action succeeded or failed.

Solution
========
Add a FAILING state to the API and ActionExecutionDTO. The POST action script can look at this new state and make a decision what to do with the action.

Alternative Solutions
=====================
1. Add a `hasFailure` flag
2. Expose the state of every exuection step (PRE, REPLACE/Native, and POST). Currently only ActionOrchestrator knows about this

Please also review the changes in:
https://rbcommons.com/s/VMTurbo/r/42563/
https://rbcommons.com/s/VMTurbo/r/42564/

Related Fixes
=============
1. Checkstyle for the files I touched in Topology Processor
2. I skipped API since the files I touched there had over 100 violations :(
3. Checkstyle for some of the the files I touched in Action Orchestrator (I didn't do all because there's over 50 :( )
4. Spotbugs enabled for Topology Processor. You may not commit something that increases the number of spotbugs violations.
5. Spotbugs enabled for API
6. Spotbugs enabled for Action Orchestrator
7. dead code removed from OperationManagerTest
8. dead code removed from Action and ActionView
9. fixed the action state diagram by adding missing edges and rotating it 90 degrees so it fits in the 100 character limit