[OM-64385]: RB:42566 Apply scaling factor consistently during market conversion
** Problem **

At Bank of America, we see an IllegalArgumentException being thrown during market analysis when we attempt to set a commodity's quantity to a very slightly negative number (approximately 1e-4).

After debugging for some time, we identified that the cause was a resize action on a Container's VCPU sold from a capacity of 1491.3689 to 0. When propagating the impact of this resize to dependent commodities, the used value on VCPU_LIMIT_QUOTA bought from the ContainerPod has a value of 1491.369. We subtract off this 1491.369 from the quantity sold by the ContainerPod, but the used value on the ContainerPod's sold commodity is only 1491.3689, and so the result of the subtraction is a negative number resulting in the IllegalArgument.

When looking in the topology diagnostics, we see that both the VCPU_LIMIT_QUOTA bought by the Container and sold by the pod are equal (a value of 540.0). However, a scalingFactor of 2.7617943704573737 is being applied and after the multiplication by scaling factor, we see this difference.

Guy has already hardened market code to prevent this sort of exception from completely halting market analysis, but there are still scenarios where this sort of floating point precision problem can lead to incorrect decisions (particularly around the capacityUpperBound and capacityLowerBound). We should fix this to try to be as consistent as possible in our handling of the scalingFactor to ensure correct market behavior.

** Solution **

Because most commodity quantities in TopologyEntityDTOs are doubles but are floats in the TraderTO's we create for the market, we have to incur some floating point precision loss during the conversion. We just need to ensure that we incur the loss in the same way for all quantities multiplied by the scaling factor, otherwise certain numerical comparisons we do (whether with 0 when setting quantities or with the commodity upper/lower bounds) may fail.

This fix updates commodity conversion logic so that we apply scaling factor logic by incurring floating point precision loss the same way for relevant fields on commodities bought and sold.

** Testing Done **
* New unit test to ensure scaling factor is applied consistently during market conversion.