[OM-63324] Changed a logging line to print OID instead of display name.
This was done to avoid potential NPE in TopologyConverter when entity
existed only in projected topology and not the source one.

Arguably I could check for `null` and print OID only when display name
isn't available but I've found use of display names in the logs only
marginally useful. There is no consistency in what is currently printed
across the platform so one of the first things I need to do is find the
mapping between OIDs and entities from the diagnostics anyway. There is
also a performance penalty from looking up the OID in the map, that can
also be avoided with a more complex pattern. All-in-all just printing
the OID seems simplest and safest at this time. If we had a logging
utility class for this kind of thing we could afford to be smarter about
what and how we print without complicating clients.

I could not reproduce the error in 8.0.4 (staging) or 8.0.2 so testing
was done by back-porting to 8.0.1. Started up Voltron and loaded
diagnostics from the issue. Run plan similar to the one in the issue but
with 3k new pods as I couldn't reproduce the problem with 300. Before
the fix I was getting NullPointerException and the plan wasn't
completing. After the fix there was no exception and the plan completed
successfully.