[OM-64431]: Allow stitching journal to print to logger on HTTP or RPC requests
** Problem **

At Bank of America, we ran into an issue where we needed to be able to get stitching journal information (to see how and why the scalingFactor on a Container was being modified) but were unable to fetch the information because running the broadcast takes too long and the HTTP request was timing out, closing the stream where we were outputting the journal log.

To ensure we are able to get the journal information even in these situations, we should also permit the output of the journal to the log in addition to returning it to the remote requestor.

** Solution **

Add a new 'output_options' flag to a StitchingJournalRequest to permit where the journal output is sent. The new default output option (LOG_AND_RETURN_TO_CALLER) will send output to both the log (ie rsyslog) and return it to the caller, whether the caller is a gRPC request or an HTTP request. There are also output options for sending only to the caller, or only to the logger if that is desired.

The LogRecorder option (used for printing journal output to the log) used to capture the entire journal and then log only at the very end of the broadcast. To ensure that we don't run the topology processor out of memory for the journal, we now flush the LogRecorder:

1. At the end of broadcast
2. At the end of every stitching operation.
3. Whenever some maximum character limit is exceeded.

** Testing Done **
* New unit tests.
* Tested in-app both the HTTP and RPC options and ensured they work appropriately.