[OM-63720] Use the action SHA1 hash rather than using composite field as a unique identifier of action
Background
==========
The way that the action identity works is that we have a composite unique index on a number of fields in db. Those fields contain the information that makes the action unique. As the primay key is not case-sensitive with current charset of db we had an issue that two actions that their data differ only in casing (like delete action with two path that only differ in case) resulted in constraint violation exception.

Feature Implemented
===================
- Add a new field in db which keeps SHA1 cache of the fields that makes the action unique. The type is field is binary(20) and has a unique index on it.
- Create a migration script which populates that field
- Remove the existing unique index on the action detail
- Change the recommendation persistence to populate that field.
- Changed in the memory cache to use this hash as the key rather than the whole action data