[OM-63799] Implement retention processing logic to prune Volume Attachment History table.
Requirements
------------
The Volume Attachment History stores the history of Volume and VM relationships. However, for Volumes which have been deleted and are no longer in the environment, there is no need to maintain any records. In this review, a mechanism whereby records for Volumes no longer discovered are being removed. The assumption is that if a Volume has not been discovered for more than 60 days, then it is no longer in the topology and has been deleted.

Summary of changes
------------------
Execute query to delete records related to volumes older than defined retention period in days (thanks to @Andy for the concise query!) from RollupProcessor's performRetentionProcessing step.

Review: https://rbcommons.com/s/VMTurbo/r/42680/