[OM-64199]: XL: Use plan source topology broadcast from topology processor, not market.
Intent
======
The topology processor broadcasts the scoped plan source topology. Components that want to process the plan source topology should listen on that topic.

Background
==========
Once upon a time the topology processor was responsible for broadcasting the plan source topology. Then, to support plan scoping, we had the topology processor broadcast the FULL plan source topology to the market. The market would do the scoping, and broadcast the plan scoped source topology from the market. All listeners were converted to listen to the Market. Finally, we moved the scoping logic to the topology processor. Today the topology processor scopes the source topology, sends it to the market, and the market sends it to the rest of the components. This is incredibly wasteful!

Fix
===
All components that care about the plan source topology are now listening to the topology processor broadcast of the plan source topology.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42464/