[OM-64177]: XL Wasted Files: Run Wasted File and Reserved Capacity analyses before regular analysis.
Intent
======
Run wasted file analysis and reserved capacity analysis earlier in the Analysis pipeline, while holding fewer resources.

Background
==========
At BofA, when the analysis starts, the Analysis object takes up about 6GB (mainly the topology). After the analysis runs, the Analysis object balloons up to 12GB. Since the Wasted File and Reserved Capacity Analyses only operates on the source entities, we can run them before the main analysis. That way we are holding the extra 6GB of memory for a smaller period of time.

Implementation
==============
This is part of an effort to limit the degree to which we need to keep the source topology in memory, so I am also refactoring the WastedFileAnalysis and ReservedCapacityAnalysis a little bit. There are now "WastedFileAnalysisEngine" and "ReservedCapacityAnalysisEngine" classes, which produce "WastedFileResults" and "ReservedCapacityResults" respectively (containing actions, and any other analysis output). The benefit is that we don't need to keep additional references to the fields that used to be in WastedFileAnalysis/ReservedCapacityResults, which removes a couple of obstacles to removing/trimming the source topology dto references after conversion.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42499