[OM-64256]: XL TP: Keep cached topology in serialized format.
Intent
======
Reduce memory usage of topology processor.

Background
==========
The "CachedTopology" in the topology processor keeps a copy of the last stitched topology so that plans can run faster without doing stitching. We currently keep a set of cloned TopologyEntityDTO.Builder objects, which occupy several GB at BofA scale.

We can get some significant savings by serializing the TopologyEntityDTO as a byte array. This will make plans marginally slower (since we need to deserialize the entities), but will shave off a couple of GB.

Implementation
==============
The "CachedTopology" now keeps TopologyEntityDTOs (instead of TopologyEntity), and a new stage follows the ConstructTopologyFromStitchingContextStage which creates TopologyEntity.Builders from TopologyEntityDTO.Builders. There is an environment property to enable/disable serialization of the cached topology. It's on by default, but we can disable it if we find that the extra CPU time is not worth the memory savings.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42493/