[OM-64807]: (RB 42776) Update the migration script for adding action hash to handle duplicate rows
Intent
======
Rewrite the migration script which replaces all action details with the hash of them to handle duplicate rows gracefully

Problem
=======
In https://rbcommons.com/s/VMTurbo/r/42594/, I introduced a migration script which replaces all the fields that keeps properties of action that makes it unique with a single field keeping the hash of it. The fields that I replaced with hashed version of them had a unique index. Therefore, I was expecting none of them have duplicate data. However, one of those fields accepted null and it seems Mariadb allows rows with duplicate data if one of the fields is NULL. Therefore, we faced migration failure on an instance that had that type of actions with null data.

Solution
========
- Since this code is not shipped to customers, I am going to update the scripts so it can handle duplicate rows and send an email with instructions how to fix failed migrations to Developers and PT.

- The solution should be removing the duplicate entries as the duplicate entries should not be there in the first place. This should not impact anyone as the oids for start actions (which were in question here) were already changing over time. There was two alternatives:
-- removing duplicates from source table, I tried different approaches this way and it seems to be inefficient.
-- create a new table insert entries into it ignore duplicates and replacing the original table with this one. This one seems to be very efficient.