[OM-63591]: UPDATED: XL Group - Create new paginated grpc call to return groups
Updated: fixed migration issue (non-backwards compatible with older
innoDB versions)

This is the first part of changes to introduce support for pagination
natively in the group component. Changes in this commit are mostly not
visible to customers yet; the only exception is case sensitive filtering
by display name.

Related story: OM63092

RB42293