[OM-64664]: RB:42839 Fix container request and quota price functions
** Problem **

With kubernetes targets added, the market was generating exceptions similar to:

[Resizer]	: Trader CONTAINER|73728583170173|nsquota-test-multi-deployments/quota-test1-7c658ff4d9-dnspc/cpu-load while resizing threw exception Requires f(4.9E-324)*f(0.9999999999999999) < 0 : Capacity 40960.0 Quantity 12296.0 Revenues 3.001953049164285E-5 Expenses 10.00000025547264

Subsequent investigation revealed that these exceptions were caused by attempting to run on the Bisection algorithm in the Resizer to find roots on a Step price function (used for request commodities) when the Bisection algorithm cannot guarantee a solution (it requires a continuous function to guarantee solutions). The reason we picked a step price function for request commodities was to ensure that request commodities do not put pricing pressure on Containers to move when request is at lower congestion (because since it is an allocation commodity, even when request is highly utilized, it usually does not have much performance impact and kubernetes will ensure request utilization remains below 100%). I also noticed that the different quota commodities did not use the same price function even though they have the same semantics in kubernetes (the REQUEST quotas used a standard price function but the LIMIT quotas used a step price function).

** Solution **

Speaking with the market team,

1. VCPU/VMEM_REQUEST: Use a standard price function with low price weight (this IS a change from the step price function we currently use).
2. VCPU/VMEM_LIMIT_QUOTA: Use a step price function (this is not a change because we already use a step price function for these commodities).
3. VCPU/VMEM_REQUEST_QUOTA: Use a step price function (this IS a change from the standard price function we currently use).

By picking a standard price function with low price weight (I think a value 0.0001 might work because it would give a unit price of 1 at 99% utilization and would match the step_below value used in the step price functions) for request commodities, requests should have very little impact on move behavior until it gets to very high congestion. However, the price weight will not affect resize behavior because the price weight appears in both the numerator and denominator of the function we are solving using bisection, so it is effectively cancelled out.

Quota commodities are bought through a separate, immovable shopping list so we do not have to worry about them affecting move behavior no matter what price function we pick for them. Vaptisis mentioned that it should be safe to pick a step price function for quota commodities because they only appear as raw materials in the function we are optimizing, and we don’t have the same restrictions there as we do for the price functions of the commodities we are resizing. We may want to return to this decision later if/when we need to start resizing quotas (and this is something we know customers want us to do).

With these changes I think we can expect:
Request commodity below 100% utilization does not affect pod moves very much (ie. Bin packing-like behavior for pod moves with respect to request).
Quota commodities don’t put resize pressure on their respective requests/limits but they prevent going above 100% quota utilization.
We should no longer see exceptions in the market when attempting to resize requests. We already use a standard price function on limits (VCPU/VMem) with a step price function on their LIMIT_QUOTA raw material and I don’t think I’ve ever seen an exception for them so I am fairly confident doing the same for request should be ok.

** Testing Done **
* Using a topology where Bisection exceptions were 100% reproducible, verified that after these changes, no more such exceptions are generated.
* Verified that after this change, overall action counts are similar (in particular with relation to requests and that some individual request actions that I scanned still make sense.