[OM-65171] Do not disable suspension and provision of application if it is a sole provider of a service in TP
https://rbcommons.com/s/VMTurbo/r/42890

Problem
-------
In TP constructTopology stage, we currently have a logic to determine if an application entity is suspendable. If the application is a **sole** provider to a service, then we disable suspension. When suspension is disabled for an application, **provision is also disabled**, which is wrong. This issue had been hidden by another bug where in ScalingApplicator stage, we blindly overwrite any existing provision setting of an entity, until that bug was fixed in [rb42742](https://rbcommons.com/s/VMTurbo/r/42742/). After [rb42742](https://rbcommons.com/s/VMTurbo/r/42742/), this issue manifests itself, i.e., when an application is a sole provider to a service, the provision is always disabled.

Solution
-------
We should remove the sole provider check for application in topology processor. In market we already have logic to prevent an application from being suspended if it is a sole provider, so the check in TP is unnecessary.