[OM-65026]: CCA - Truncate analysis window end time based on analysis interval
rb: 42871

**Background**
In CCA, the analysis will query back X number of days in selecting the demand to analyze. It will then truncate any demand that stretches across the start of the analysis boundary. For example, if the analysis window starts at 1 PM and there's a VM entry from noon until 2 PM, the demand retrieval stage will select the entry and then update the start time to 1 PM (to stay within the analysis window).

**Bug Description**
The truncation of demand across the analysis boundary was performed for the start of the analysis window, but not for the end of it. This code review adds truncation to the end of the analysis window.