[OM-65120]: RB-43206: ActionStream Book keeping schema
Intent
============
Add table to store book keeping so that we can use that data in that table to recover from component restart or failure.

Background
============
We send actions to workflows configured by policy only when the action was first generated and has not cleared. We also need to notify the workflow when the action has cleared.

Problem
============
Need to remember which actions we sent in the case of restart so we do not send the again.
Need to remember when an action disappared so we can evaluate the cleared criteria.
We do not need to remember which actions have not been delivered yet because that will be handled by the internal kafka used for communication between AO and TP.

Solution
============
1. an action_recommendation_id, workflow_id pair
2. can quickly look up state if needed by action_recommendation_id, workflow_id
3. a cleared_timestamp set when we first notice that an action is no longer seen by the market

We implement the schema in this way to minimize the number of writes needed. Writes are minimized because we only write changes every market cycle instead of all actions.

Related Fixes
============
Removed unused checksum from db-migration-checksums.properties