[OM-64527]: CCA - Fix resolving of RI inventory when all RIs are excluded from the plan scope
rb: 42884

**Background**
The cloud commitment resolver is responsible for resolving the cloud commitments, based on the OIDs supplied by the plan-orchestrator in requesting the analysis. In querying the RI bought store, the commitment resolver did not check whether the OID list is empty, causing it to return all RIs in inventory when the intent is to ignore all RIs in the plan scope.

**Code fix**
The fix is straightforward - simply check whether the OID list is empty.