[OM-64717]: Multiple Policies Conflict Resolution RB:42919
Background
===============

When two policies share the same automation mode, we show both of them in the policies section for a group.

Problem
=========

Not picking a winning policy makes it hard for a customer to determine which policy will be chosen when automating actions.

Solution
=========

The reason to introduce this feature was to show all policies that may have a chance of winning was because of our requirements with RTA. For the RTA feature we were asked to support execution windows in settings policies.
Customer scould have muliple execution windows per each (actionType - actionMode) pair.
However, the change was affecting policies without schedules. This change only forces the multiple policies to be chosen when schedules are present, else we can have one winning policy chosen.