[OM-65253]: CCA - Add ImmutableTimeSeries and fix flexible demand transformation for multiple cloud tier records
**Bug Background**
The `DemandTimeSeries` incorrectly had the `@Auxiliary` annotation added to the time series attribute. In cases in which we had multiple flexible instance sizes for a single entity, this was improperly treating two `DemandTimeSeries` as equal. The fix is to swap the `@Auxiliary` attribute with `@Redacted`, given the intent in using the auxiliary annotation was to hide the time series during logging by default.

**ImmutableTimeSeries**
This commit replaces the `TimeSeries` implementation with a cleaner `ImmutableTimeSeries`, which is a delegated wrapper around `ImmutableSortedSet`. This change faciliated removing many of the helper methods to create `TimeSeries`, as well as cleaning up the `TimeSeriesEncoding` class, which will now only be applied to immutable classes with `TimeSeries` attributes.