[OM-65206]: CCA - Skip empty demand and handle duplicate start time
rb: 42925

**Background**
CCA assumes all end timestamps for demand are inclusive, meaning no VM should have an entries where the end time of one entry is equal to or overlaps with the start time of another. However, with PT demand injection during testing, this assumption about the demand was broken, which caused CCA to drop one of the entries. The cause for the drop is the use of `TimeSeries`, in which was treating two entries with the same start time as equal. While this issue will not happen under normal operation, it is useful to handle for PT testing.

**Solution**
There are two fixes for this:
1. Drop any empty entries i.e. entries in which start time == end time. This is useful to avoid spurious processing in later stages.
2. In `TimeSeries`, add a tie-breaker for `TimeInterval` and `TimeSeriesData`