[OM-65317]: Support ConsistentScaling on heterogeneous providers RB:43198
** Problem **

We model a group of entities replicated from some shared definition using what we call a “Consistent Scaling Group”. Because they are replicas of each other, members of a consistent scaling group must share certain features in common. The most important of these is that commodities resized across the scaling group must share the same capacity.

We have problems with consistent scaling CPU resources when the suppliers for the scaling group members have different CPU speeds. CPU resource values (capacity, used, etc.) are all normalized to be of the same unit (normalized MHz). CPU resources for scaling groups are usually defined in terms of cores (for VMs) or millicores (for containers) and not in MHz (and certainly not Turbo normalized MHz). CPU resource capacities for a scaling group are required to be consistent in this original group unit of cores/millicores. If all group members are running on providers with the same Turbo normalized CPU speeds, there is no problem - the originally consistent capacities in cores/millicores will be consistent in the normalized units as well. However, if, for example, some group members are running on providers with fast CPU speeds and some members are running on providers with slow CPU speeds, because the market sees capacities in normalized units, the market sees inconsistent capacities for the group members even though they are consistent in their original unit.

The market responds to this inconsistency by trying to make all group members consistent. It picks the highest capacity from all group members and generates a resize action to take all group members to this capacity. In our example above, this would mean the market would generate a resize action to scale the group from 200MHz to 300MHz in order to make all group members consistent. Unfortunately, this is a nonsense action because the capacities were actually already consistent, just not in the units the market was using. A similar situation happened at a customer where for example where we resized up a container at 2% utilization because of this inconsistency.

** Solution **

We introduce a new double value (called "consistent_scaling_factor") on AnalysisSettings that can be used to convert from the normalized economy units (ie MHz) where group members might not be consistent back to group-local units (ie millicores/cores) where group members should be consistent. This consistent_scaling_factor is applied in the market in the ConsistentResizer. This review includes the following changes to ensure the correct scalingFactor and consistentScalingFactor are set on the appropriate entities involved in resizing decisions.

1. CpuScalingFactorPostStitchingOperation (modified) - This operation sets the ScalingFactor used to normalize MHz values across CPU architectures where MHz is not a good measure of performance. We should now scale requests by the ScalingFactor as well to ensure the same CSF used for limits can also be applied to requests. The same goes for request quotas. We also now take the maximum ScalingFactor for a WorkloadController and Namespace rather than an arbitrary ScalingFactor. Note that this change to WorkloadController and Namespace only affects quota commodities and it still leaves a feature gap around quotas to be addressed later.
2. CpuConsistentScalingFactorPostStitchingOperation (new) - This is a new PostStitchingOperation that sets ConsistentScalingFactor on:
   2a. VirtualMachines discovered by cloud native targets - We do not set CSF on VMs NOT discovered by cloud native targets so their resize behavior is not affected. CSF on a VM is set such that normalizedMHz * CSF => millicores
   2b. Namespaces - Set the CSF on namespaces. Note that we do not set CSF on containers here because it will be done by the EphemeralEntityEditor. We do not set CSF on pods, or workloadControllers because they are resellers and so their commodities are not actually involved in the resize decisions.
3. EphemeralEntityEditor (modified) - Set the CSF on containers that are resizable. The container CSF is set to be equal to the CSF of the provider node. Updated the summary of changes to include the number of containers where the CSF was set.

Note that VMs hosting containers are not resizeable so this review has no impact on VM resize behavior.

Note also that because a namespace is spread across multiple nodes, kubeturbo converts its quota commodities from MHz -> millicores by using the average speed of all nodes in the entire container platform cluster where the namespace is running. In order to allow inverting this computation to go from MHz back to millicores, kubeturbo now sends the average_node_cpu_frequency which is now also included on the TopologyEntityDTO for namespaces.

We also introduce a feature flag (enableConsistentScalingOnHeterogeneousProviders) in the TopologyProcessor to control whether we will generate resize actions when a scaling group is spread across providers with different CPU speeds. If false, we will disable resize on the Containers in such a scaling group (the prior behavior). To enable this flag, add the following property to the CR file:

spec:
...
  properties:
    global:
      enableConsistentScalingOnHeterogeneousProviders: true

applying it and restarting the TP pod.

** Testing Done **

* Duplicated all ConsistentResizer tests to also test behavior in a heterogeneous provider scenario (where members of a consistent scaling group have different ConsistentScalingFactors).
* Ensured all market tests still pass.
* New unit tests for ConsistentScalingQuantity class.
* Tested in-app to ensure that when containers in a consistent scaling group are spread across a set of heterogeneous providers, we can still get a sensible resize action.
* Verified we do not get scaling actions for scaling groups on heterogeneous providers when the enableConsistentScalingOnHeterogeneousProviders flag is disabled.
* Executed an action in the lab on a heterogeneous cluster (dc11 target) and saw the deployment correctly updated.
* Verified VMs that are not part of the kubernetes cluster still get the same resize actions.