[OM-64640]: XL Report: Refactor actions schema. Add rotating "Pending Actions" table.
Intent
======
Simplify action schema design and support action pruning.

Background
==========
We have had issues with the action and action spec tables growing unboundedly at larger customers, and wanted to prune the table. When looking at the schema we decided to simplify the model - namely, to keep the "pending" actions in a single table. This allows us to rotate the actions (write snapshots to a new table, swap that new table with the old table, and drop the old table) which is a simple approach that should scale relatively well. In the future we can use hashes + targetted updates.

Implementation
==============
Most of the changes are in the V1.11 migration, and the **ReportingActionWriter** and **ActionConverter**. It's simple stuff, because the new schema is significantly simpler than the old one.

The vm_rightsizing_actions.yaml file contains the updated VM query for the report.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42958