[OM-65047] SMA - Support Operating System License Model based cost calculation.
Requirements
------------
In Azure, Windows VMs with Azure Hybrid use Benefit (AHUB) enabled are subject only to the Compute portion of the on demand rates. This is equal to the Linux on demand rates. This is because if the VM has AHUB enabled, it implies that they are using an already purchased on-prem Windows License on that VM and hence there's no need to charge for Windows License in the on demand rate.

Further, if such VMs are covered by an RI, they should also not be subjected to the Reserved Windows License rate (in SMA parlance, the discounted cost) for the same reason as stated above, which is that the Windows License being used is already paid for.

Due to the current lack of support of the SMA cost calculation to consider License Model, actions are seen which move VMs to sizes which are cheaper for Windows rates, however these sizes are more expensive for Linux rates and cause an investment action (incorrectly).

As a part of this review, add support to SMA cost calculation to consider License Model as a parameter. If the License Model is AHUB, then use the Linux OS Type to look up the on demand rate. Also, if the License Model is AHUB, then return 0 for the discounted cost.

Summary of Changes
------------------
1. Introduced attribute operatingSystemLicenseModel to SMAVirtualMachine, value comes from LicenseModel field in VM info (of TopologyEntityDTO).
2. Introduced CostContext to hold business account, region and os License Model - holding fields used for price lookup.
3. Modified SMATemplate methods getNetCost, getOnDemandTotalCost and getDiscountedTotalCost to factor in the License model.
4. Refactored callers of the above 3 methods to pass in the new parameter.

Review: https://rbcommons.com/s/VMTurbo/r/43052/