[OM-65220]: XL AO: Avoid exception if source entity of move is not present in snapshot.
Issue
=====
Action plans not being fully processed at Citi.

Background
==========
At Citi we saw that a source entity id was not present in the topology snapshot. This was causing a runtime exceptions when refreshing actions in LiveActions. The runtime exception was short-circuiting a lot of the action post-processing.

Note - we don't know why this entity was not present.

Fix
===
The fix is twofold:
1) Check to make sure the source entity is present in the ActionDescriptionBuilder.
2) Catch runtime exceptions in RefreshActions to prevent an error in one action from stopping processing of all actions.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42998/