[OM-65353]: XL History: Reduce possibility of ThrottlingCompletingExecutor lockups.
Issue
=====
Plan and realtime ingestion intermittently hanging.

Background
==========
I noticed that the following log line appears in the diags:
[ThrottlingCompletingExecutor]  : No handler registered for completion of future: java.util.concurrent.FutureTask@2d56aef0[Completed normally]

This led me to believe that the executor service inside ThrottlingCompletingExecutor is out of sync with the "pendingExecutions" map - e.g. the task is completing before the "pendingExecution" is updated. This doesn't seem very likely, because the task involves database access, but looking at the code it seems like the only way that the future can be absent from the map.

If that's true, then the future that will be put into the "pendingExecutions" will never get taken out, and any attempts to "drain" or "close" the ThrottlingCompletingExecutor will hang forever.

Fix
===
My attempted fix is to simplify the workflow and call the handler (wrapped in a future to represent exceptions) directly from the threadpool. This eliminates the need for a "pendingExecutions" map. I'm not 100% convinced this fixes the issue, but I'm not able to reproduce it locally, and this will definitely remove one way of getting into the "stuck" situation. This also gets rid of the completer thread.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/43031/