[OM-64730] Deactivate commodities with autoset capacity
https://rbcommons.com/s/VMTurbo/r/42870

Problem
========
When SLO is not enabled in the policy, but used SLO values are sent from the probe, post stitching operation will autoset a capacity for the commodity. The autoset value is calculated based on historical used value. When provision/suspension is enabled for the application, the RoI are computed based on the autoset capacity, which does not make sense.

Solution
========
The proposed solution is to **not** send the commodity with autoset capacity to the market. To do so, I am reusing the `active` attribute of the commodity. Inactive commodities will not be sent to the market.

Implementation
========
If a commodity needs autoset capacity (i.e, SLO is not enabled), set the commodity **inactive**. We will deactivate both sold commodities from the sellers and the corresponding bought commodities from the consumers inside `SetAutoSetCommodityCapacityPostStitchingOperation`. The performance impact is limited because we are only doing this for Transaction and ResponseTime commodities for those entities coming from APM probes.